refactor(vue): remold vue package
