+++ packages/vue/src/hooks/index.ts
@@ -1,2 +1,2 @@
-export * from './event-lock'
-export * from './form'
+export * from './use-event-lock'
+export * from './use-form'
