+++ packages/vue/src/hooks/use-event-lock.ts
